Add bootstrap
In order to easily create well-looking pages, using a css framework is helpful. For now, we will test to use bootstrap as css framework (might be changed later). Bootstrap was installed via npm since this allows us to adapt it to our needs, e.g. to change the primary colors to our presentable- / hytea-orange.