Add option to add content for AudienceBeginning
The user should be able to add what the audience already knows to AudienceStore.knowledgeBeginning. For reusability reasons, a SubpointAdder (textfield + button "Add") was created which emits the content of the textfield to the parent so that the parent (in this case AudienceBeginningView) can decide what to do with the content to add. In this case it will be added to AudienceStore.knowledgeBeginning