Add option to remove elements from ContentBox
If users e.g. decide that they want to delete on point they already added, they can now just click on the x and it is removed. Since the ContentBox should be flexible and should later (e.g. in main) also only display content without the option to delete it, this is optional (configurable with hasRemoveOption). For AudienceBeginningView, the x now removes the corresponding content from AudienceStore.knowledgeBeginning.