Improve and unify site layout
Moved all containers as far "outside" as possible, so that the layout is as unified as possible and that we don't need to re-define all of it in each of the sub-views.
Furthermore, the layout was unified, centered and margins, paddings etc. adapted to have a clearer more unified layout.