Add SubstepsViewer
The IntroductionView will consist of multiple substeps (i.e. multiple texts shown to the user after each other). However, all of them are part of preparing the introduction so we don't want to create separate views for all of the, especially since most are just texts + button. Therefore, IntroductionView will feed the SubstepsViewer the information for the current substep and then update the content if the substep changes.