Add keyboard control for SubstepsViewer
Clicking on "next" all the time gets cumbersome quickly. Therefore, the possibility to navigate by pressing keys was added. The benefit of the KeyboardListener component is that it adds the keydown-handler globally for the time it is being used so that the user does not e.g. need to click on the component with the keydown event first but can just directly press the key and the correct action is triggered.