Save chosen IntroType
The chosen type of introduction (e.g. "joke", "story", etc.) needs to be saved for later usage (e.g. for when defining the details of the introduction). Also, later on we need to be able to save and read this from e.g. the server and/or json-files. Now it is stored in a Pinia-Store which is accessible easily from all components.