Move centering container inside Views
If the centering container is in MessageCompositionView, this means no of the other views can display something outside of the center. For when the user is creating the end, however, we now plan to also show things on the left and right. Therefore, now the view for every step decides for itself whether the content is centered or not. To make this a bit easier, the component CenteringContainer is providing exactly this centering functionality. If e.g. the width needs to be adjusted, it can be adjusted there for all views using it.