Make elements in ContentBox draggable
If users want to reorder the content of their presentation, it would be great if they could just drag and drop the content elements into a different order. This commit enables a drag and drop behavior like this. Previously, users would need to remove and readd the content in a different order.