Remove arrow buttons from ChaptersViewer
The arrow buttons were irritating (i.e. "When should I use these arrow buttons and when the 'Next' and 'Back' on the bottom?"). Additionally, they were only used in the IntroductionView and not in the other preparation steps. In the other preparation steps we provided only keyboard navigation. Therefore, the arrow buttons were removed to unify the different views.