Update Readme
Right now, the composition API and the options API are both used in components and views. To unify this, in the future all components and views should only use the components API. Therefore, a corresponding hint was added in the Readme. Since the existing code will to large degrees be replaced because of the new design, the remaining components will be refactored to use the composition API during the replacement process.

Furthermore, there was still a reference to vuese-documentation in the folder overview of the readme. Since vuese is not used anymore, this reference was replaced with the reference to the folder containing the "new" documentation with vue styleguidist.