Move files into MC subfolder
As we will put the message composition and the rehearsal part of the application into one repository in order to be able to use components together, the folders "views" and "stores" will have one subfolder each for the message composition part (abbreviated as "MC") and the rehearsal part. This commit moves the existing message composition files into the corresponding subfolders and adapts the corresponding filepaths.