Add components for design v2
V2 of the design will consist of a chat-like interface where the parrot sends chat messages to the user (handled by TextBlocks.vue) with interactive blocks in between for user input (handled by UserInput.vue). The progress of the user is saved in ChapterProgressStore.js and this store's initialization as well as the connection to the KeyboardListener is managed by ChatInterface.vue.