Simplify TextBlocks.vue
Previously, each element in the texts array given to TextBlocks.vue needed to consist of the text as well as an id. This was due to some relicts from design v1 (there text blocks could be interrupted by showing something in the side pane). In the design v2, however, we don't need the functionality to show things in a side pane in between TextBlocks anymore. Therefore, we could simplify the texts-array in TextBlocks.vue so that the user does not need to manually write the id for each text anymore since it will be the same as the position in the array.