Simplify transition between steps
Before, one would need to press the "Next" button on the bottom to go from one step to another. This was unintuitive since it breaks the flow of the application. With this commit, ChapterProgressStore checks inside it's next function whether the current step is completed and if yes starts the next step.