Remove buttons in Footer
The buttons "Next" and "Back" which have been on the bottom of each page so far have been removed. They were useful when the navigation was not done with arrows and also for debugging, but now they are not needed anymore. Additionally, they might be confusing for the user as e.g. a click on "next" would not show the next text message but the next big step.