Add auto-scroll to bottom of page
When the user triggers next (e.g. by pressing the right arrow) and a new text message is displayed, until now if the page was full, the user would need to manually scroll down to the new text message. With this commit, whenever a new item is added the page will automatically scroll to the bottom so that the user can immediately see the newly added e.g. text messge.