Add option to use a custom button in UserInput
In the IntroductionView we will have a custom button (because this button has additional functionality, e.g. it saves the IntroductionType). Other views (e.g. rehearsal) might even have other mechanisms than a button (e.g. that displaying the next part might be triggered when the replay is over). To increase the flexibility of using UserInput, the button is in this commit made optional and the stopWaiting can be triggered by emitting stopWaiting inside the slot.