Add CenteringCol5.vue
There were multiple instances where content should be centered with a width of col-5. Therefore, the component CenteringCol5 was created to take care of this and to make later changes (e.g. potentially from width 5 to 6) easier since they only need to be done in one place. Furthermore, it's easier to reuse this now.