Remove unneeded comments in ProgressStore
The steps "Main-2" and "Main-3" have been in the original plan for the software but were removed some time ago. To keep the option to re-add them easily they were kept in comments. Now we know that we will not re-add them so we can safely remove them to decrease visual clutter.