Add option to not wait in UserInput
In the EndView we want to display some information (recap of introduction plan) that the user should just read and not (yet) act upon. Therefore, it is not needed to wait for the user to finish and click a specific button, but it is fine if the user just uses the normal next-function (e.g. right arrow). This commit introduces the corresponding functionality in UserInput.vue + makes use of it in EndView.