First draft of WelcomeView
The WelcomeView should welcome the user and introduce the user to the project. This commit adds a first (incomplete) draft of this WelcomeView that needs to be extended in the future.
Since this change adds a chapter and therefore adapts the order in which the steps should be displayed, changes in e.g. ProgressBar and ProgressStore were needed as well.