Bugfix: previous triggered when editing text
This commit fixes a bug, where if the user's cursor is inside a textfield, the user might use the arrows to navigate inside the textfield (i.e. to go a few char to the left) and the "previous"-action was triggered accidentially.
Now, the previous-action is locked while the user's cursor is inside a textfield.