Bugfix ContentBox: automatic line-breaks
Before, if the texts the user wrote (e.g. about the audience's previous knowledge) was too long, the added text was visually sticking out of the content box. With this commit, automatic line-breaks are added if the content is too long so that it (visually) stays inside the parent element.