Improve WelcomeView (wait if "No" chosen)
Before, if the user chose "No" when being asked if they want to try the software, the user could still go to the next step. However, if the user chose to not wanting to try the software, the software should stop displaying further instructions etc. This behaviour has been implemented in that commit by setting waiting to true after the user chose "No" and the parrot displayed its response message.