Bugfix "previous" (add. speech bubble)
When the previous step ended with a TextBlock and the user went one step back, additional empty speech bubble(s) have been shown at the bottom of the previous step. This was due to the currentElementID being set to an array containing the length of the last chapter: Now, it is set to the last ElementID of the last chapter.