Add tracking for next, previous, jump actions
In our experiments, we want to figure out how much time people spend with each part of the preparation and how they navigate through the program. Therefore, we need to track when they go to the next element, to the previous element and when they do jumps. This tracking has been implemented in the current commit.