Add option to write the pitch's topic
As a first simple task, users should type in the topic of their pitch.
To enable them to change that after typing it in first, SubpointAdder was adapted in a way that if the prop keepDisplayingContent is set to true, the content is not deleted from the input field after clicking on the button and that the button's name is afterwards changed to "Change" (instead of "Add").