Improve navigation (click, arrowUp, arrowDown)
With this commit, a click somewhere on the page triggers a next event as people might try to click when they don't know how to navigate this application.
Since the conversation with the parrot is horizontal, using the keys arrowUp and arrowDown as additional navigation options, might be more intuitive.
All these navigation options have been additionally added in the current commit.

Since the "next" event is now triggered on each click and "unsetWaiting" is triggered by a click on a button, we don't need the explicit call of "next" in "unsetWaiting" anymore. If we would have kept it, the next two stepps would have been revealed since "unsetWaiting" is called before the click-EventHandler checks for a click and (if a click was recognized) triggers "next".