Change ProgressBar circles to arrows
To better visually illustrate that the steps are steps that follow each other and to visually separate them from "normal" buttons, the progress bar icons are now arrows instead of circles.
There are three types of arrows: before, middle and after. For each arrow type there is a outline and a filled version as well as a template (created in inkscape) which can be used to make changes.