Remove "I'm done"-buttons for Recap
If the user just needs to recap something (e.g. what their audience should learn) it's basically just reading and not adding things, so we don't need them to click on an "I'm done" button or have some waiting behaviour but can allow users to proceed as if this was simply another message of the parrot. This commit removes the "I'm done"-button and waiting behaviour from all parts where users only need to recap something.ö