Add asking for & saving the researchID
To be able to align the collected data of the participant & the survey data, we will use a researchID. Each participant has an individual researchID (an animal) that is entered into the software as well as into the survey. This commit adds a question for asking the participant to provide their researchID and saves it in the presentationPlanStore.