Bugfix WelcomeView: re-deciding for "No"
Previously, if the user first chose "Yes" in the WelcomeView, continued a bit and then redecided for "No", the messages displayed after choosing "Yes" (e.g. asking the user to enter their topic) was still shown. The changes in this commit ensures that these messages are not shown anymore when the user re-decides for "No".