Bugfix WelcomeView: display "Yes" when going back
Previously, if the user went to the next step and then back to the WelcomeView, it looked as if the user chose "No" even though originally "Yes" has been chosen. Since the program is stopped when the user chooses "No", we can safely assume that if a user already visited the next step they chose "Yes" so this commit makes "Yes" the standard choice and by this fixes the previously described bug.