Allow html formatting in speech bubbles
We plan to add a running example in the  message composition part. To be able to visually distinguish it from the rest, we plan to e.g. display parts of the text in italics. Before this commit, it wasn't possible to give html code to TextBlocks.vue. Now it is possible.

Hint: when we want to display content the user entered in speech bubbles at some later point in time, we would need to escape it properly to avoid security issues.