Show summary before the questionnaire
During the test run, participants reported that the software was missing an overview of their full presentation plan since they did not realize that this overview will be shown after they finished the survey. Therefore, they suggested to first show the summary and then the questionnaire. This recommendation is implemented in the current commit.