Fix "data lost on reload" bug
If the user reloaded the page (e.g. also by clicking "back" in the browser), previously all their data was lost. This shouldn't be the case so now the plugin "pinia-plugin-persistedstate" is used to persist all data stored in one of the stores. The data is stored as long as the tab that the message composition software is opened in is not closed (`storage: sessionStorage`). In case the data should be saved for an unlimited amount of time, we would need to use `storage: localStorage`. However, since users may want to create multiple presentations with the software, I decided to go for sessionStorage (=data being deleted when closing the tab). If users want to save the data, they should use the "save"-option (which will hopefully be coupled with an "import"-option soon).