Prepare repository for publishing on GitHub
Adapt the README to talk about SPEAKS (which this repository contains)
and not Presentable (which is a combination of SPEAKS and other
modules). Additionally, SPEAKS should be published on GitHub pages so
that others can test it. Therefore, the needed files have been added and
changes to the vite configuration have been made.